* refactoring

+++ PVT.Q1.2017.Shop/PVT.Q1.2017.Shop.Tests/DataBaseTest.cs
@@ -1,24 +1,21 @@
-﻿using System;
-using System.Text;
-using System.Collections.Generic;
-using Microsoft.VisualStudio.TestTools.UnitTesting;
-using Shop.DAL.Context;
-
-namespace PVT.Q1._2017.Shop.Tests
+﻿namespace PVT.Q1._2017.Shop.Tests
 {
+    #region
+
+    using global::Shop.DAL.Context;
+
+    using Microsoft.VisualStudio.TestTools.UnitTesting;
+
+    #endregion
+
     /// <summary>
     /// Summary description for DataBaseTest
     /// </summary>
     [TestClass]
     public class DataBaseTest
     {
-        public DataBaseTest()
-        {
-            //
-            // TODO: Add constructor logic here
-            //
-        }
-
+        /// <summary>
+        /// </summary>
         private TestContext testContextInstance;
 
         /// <summary>
@@ -29,48 +26,40 @@ namespace PVT.Q1._2017.Shop.Tests
         {
             get
             {
-                return testContextInstance;
+                return this.testContextInstance;
             }
+
             set
             {
-                testContextInstance = value;
+                this.testContextInstance = value;
             }
         }
 
-        #region Additional test attributes
-        //
-        // You can use the following additional attributes as you write your tests:
-        //
-        // Use ClassInitialize to run code before running the first test in the class
-        // [ClassInitialize()]
-        // public static void MyClassInitialize(TestContext testContext) { }
-        //
-        // Use ClassCleanup to run code after all tests in a class have run
-        // [ClassCleanup()]
-        // public static void MyClassCleanup() { }
-        //
-        // Use TestInitialize to run code before running each test 
-        // [TestInitialize()]
-        // public void MyTestInitialize() { }
-        //
-        // Use TestCleanup to run code after each test has run
-        // [TestCleanup()]
-        // public void MyTestCleanup() { }
-        //
-        #endregion
-
+        /// <summary>
+        /// </summary>
         [TestMethod]
         public void CreateDataBaseTest()
         {
             using (var context = new ShopContext())
             {
-                if (context.Database.Exists())
-                {
-                    context.Database.Delete();
-                }
+                if (context.Database.Exists()) context.Database.Delete();
 
                 context.Database.Create();
             }
         }
+
+        // You can use the following additional attributes as you write your tests:
+        // Use ClassInitialize to run code before running the first test in the class
+        // [ClassInitialize()]
+        // public static void MyClassInitialize(TestContext testContext) { }
+        // Use ClassCleanup to run code after all tests in a class have run
+        // [ClassCleanup()]
+        // public static void MyClassCleanup() { }
+        // Use TestInitialize to run code before running each test 
+        // [TestInitialize()]
+        // public void MyTestInitialize() { }
+        // Use TestCleanup to run code after each test has run
+        // [TestCleanup()]
+        // public void MyTestCleanup() { }
     }
 }
